Added all nfc adapter logic to the BaseActivity class, making it abstract as well for optional utilization of nfc listening from the activities that need it, resulting in the rest of the activities to ignore nfc intents properly.
