--- .idea/libraries/Gradle__androidx_activity_activity_1_2_3_aar.xml
@@ -1,18 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: androidx.activity:activity:1.2.3@aar">
-    <ANNOTATIONS>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/25a17d4fe5317469c758491f3c3b7067/transformed/jetified-activity-1.2.3/annotations.zip!/" />
-    </ANNOTATIONS>
-    <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/1a40e83ae200e4ae45fecb800d90d27b/jetified-activity-1.2.3/res" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-2/files-2.1/d128facc77236ff5d81a90815740df63/jetified-activity-1.2.3/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/25a17d4fe5317469c758491f3c3b7067/transformed/jetified-activity-1.2.3/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/25a17d4fe5317469c758491f3c3b7067/transformed/jetified-activity-1.2.3/AndroidManifest.xml" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/25a17d4fe5317469c758491f3c3b7067/transformed/jetified-activity-1.2.3/res" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.activity/activity/1.2.3/90a39fdeb6c8aa134518e9fc089fd2625a0329b0/activity-1.2.3-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
