--- .idea/libraries/Gradle__androidx_appcompat_appcompat_1_3_0_aar.xml
@@ -1,16 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: androidx.appcompat:appcompat:1.3.0@aar">
-    <ANNOTATIONS>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/7d5df6fa97a87c8b992117dcca872bcb/transformed/appcompat-1.3.0/annotations.zip!/" />
-    </ANNOTATIONS>
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/7d5df6fa97a87c8b992117dcca872bcb/transformed/appcompat-1.3.0/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/7d5df6fa97a87c8b992117dcca872bcb/transformed/appcompat-1.3.0/AndroidManifest.xml" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/7d5df6fa97a87c8b992117dcca872bcb/transformed/appcompat-1.3.0/res" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.appcompat/appcompat/1.3.0/45a346a05e7ff27a01c59452379c5939cbb53c03/appcompat-1.3.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
