--- .idea/libraries/Gradle__androidx_appcompat_appcompat_resources_1_3_0_aar.xml
@@ -1,13 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: androidx.appcompat:appcompat-resources:1.3.0@aar">
-    <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/e8582ce712d10b8b128cd4a513c5a69e/transformed/jetified-appcompat-resources-1.3.0/AndroidManifest.xml" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/e8582ce712d10b8b128cd4a513c5a69e/transformed/jetified-appcompat-resources-1.3.0/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/e8582ce712d10b8b128cd4a513c5a69e/transformed/jetified-appcompat-resources-1.3.0/jars/classes.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.appcompat/appcompat-resources/1.3.0/fa1169a9b404f7e905756330af6d13e27dcee50e/appcompat-resources-1.3.0-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
