+++ .idea/libraries/Gradle__androidx_appcompat_appcompat_resources_1_3_1_aar.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Gradle: androidx.appcompat:appcompat-resources:1.3.1@aar">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/1a13971d06ea804543051987033e3152/transformed/jetified-appcompat-resources-1.3.1/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/1a13971d06ea804543051987033e3152/transformed/jetified-appcompat-resources-1.3.1/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/1a13971d06ea804543051987033e3152/transformed/jetified-appcompat-resources-1.3.1/AndroidManifest.xml" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.appcompat/appcompat-resources/1.3.1/fa1169a9b404f7e905756330af6d13e27dcee50e/appcompat-resources-1.3.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
