--- .idea/libraries/Gradle__androidx_constraintlayout_constraintlayout_2_0_4_aar.xml
@@ -1,11 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: androidx.constraintlayout:constraintlayout:2.0.4@aar">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-3/8d98eeb656d373d523b554e09fbb1eca/transformed/constraintlayout-2.0.4/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/8d98eeb656d373d523b554e09fbb1eca/transformed/constraintlayout-2.0.4/res" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-3/8d98eeb656d373d523b554e09fbb1eca/transformed/constraintlayout-2.0.4/AndroidManifest.xml" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
