--- .idea/libraries/Gradle__androidx_constraintlayout_constraintlayout_solver_2_0_4.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: androidx.constraintlayout:constraintlayout-solver:2.0.4">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/androidx.constraintlayout/constraintlayout-solver/2.0.4/1f001d7db280a89a6c26b26a66eb064bb6d5efeb/constraintlayout-solver-2.0.4.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
