+++ app/src/main/java/com/dangerousthings/nfc/pages/BaseActivity.java
@@ -1,7 +1,11 @@
 package com.dangerousthings.nfc.pages;
 
+import android.app.PendingIntent;
 import android.content.Context;
+import android.content.Intent;
+import android.content.IntentFilter;
 import android.content.SharedPreferences;
+import android.nfc.NfcAdapter;
 import android.os.Bundle;
 import android.view.View;
 import android.view.Window;
@@ -16,8 +20,12 @@ import com.dangerousthings.nfc.utilities.ColorUtils;
  * A base activity that handles changing the StatusBar color, activating the current theme, and back press animations
  */
 
-public class BaseActivity extends AppCompatActivity
+public abstract class BaseActivity extends AppCompatActivity
 {
+    IntentFilter[] _intentFilterArray;
+    PendingIntent _pendingIntent;
+    NfcAdapter _adapter;
+
     @Override
     protected void onCreate(Bundle savedInstanceState)
     {
@@ -36,6 +44,8 @@ public class BaseActivity extends AppCompatActivity
         }
 
         super.onCreate(savedInstanceState);
+
+        nfcPrimer();
     }
 
     @Override
@@ -43,5 +53,52 @@ public class BaseActivity extends AppCompatActivity
     {
         super.onBackPressed();
         overridePendingTransition(0, 0);
+    }
+
+    private void nfcPrimer()
+    {
+        _adapter = NfcAdapter.getDefaultAdapter(this);
+
+        _pendingIntent = PendingIntent.getActivity(this, 0, new Intent(this, getClass()).addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP), 0);
+
+        IntentFilter ndef = new IntentFilter(NfcAdapter.ACTION_NDEF_DISCOVERED);
+        try
+        {
+            //add any additional NDEF related mimetypes here
+            ndef.addDataType("*/*");
+        }
+        catch(IntentFilter.MalformedMimeTypeException e)
+        {
+            throw new RuntimeException("fail", e);
+        }
+
+        _intentFilterArray = new IntentFilter[] {ndef};
+
+        handleActionDiscovered(this.getIntent());
+    }
+
+    void handleActionDiscovered(Intent intent)
+    {
+    }
+
+    @Override
+    public void onPause()
+    {
+        super.onPause();
+        _adapter.disableForegroundDispatch(this);
+    }
+
+    @Override
+    public void onResume()
+    {
+        super.onResume();
+        _adapter.enableForegroundDispatch(this, _pendingIntent, _intentFilterArray, null);
+    }
+
+    @Override
+    public void onNewIntent(Intent intent)
+    {
+        super.onNewIntent(intent);
+        handleActionDiscovered(intent);
     }
 }
