+++ app/src/main/res/layout/item_ndef_record.xml
@@ -37,7 +37,6 @@
             android:layout_marginTop="8dp"
             android:layout_marginBottom="8dp"
             android:fontFamily="@font/proxima_nova_bold"
-            android:text="MimeType"
             android:textColor="?attr/colorSecondary"
             android:textSize="20sp"
             app:layout_constraintBottom_toBottomOf="parent"
@@ -63,7 +62,6 @@
             android:layout_marginStart="16dp"
             android:layout_marginEnd="8dp"
             android:fontFamily="@font/proxima_nova_regular"
-            android:text="458 Bytes"
             android:textColor="?attr/colorSecondary"
             android:textSize="20sp"
             app:layout_constraintBottom_toBottomOf="parent"
